Fixed issues where tests were testing against notes/videos/tags instead of Optionals. Updated mockito verion in pom.xml to fix conflicts between mockito and powermock. Removed duplicate junit dependencies in pom.xml. Moved ServiceTest files into the services test folder.
